Cambios en los DTOs: los campos tipo String id fueron cambiados a Long

+++ src/main/java/com/avanzada/alojamientos/DTO/ImageDTO.java
@@ -3,7 +3,7 @@ package com.avanzada.alojamientos.DTO;
 import java.time.LocalDateTime;
 
 public record ImageDTO(
-        Integer id,
+        long id,
         String url,
         String thumbnailUrl,
         Boolean isPrimary,
